Put data in endpoint /search
